requirements.txt
pelican==3.5.0
Markdown==2.6.2
polib==1.0.6
#iso3166==0.7
python-slugify==1.1.2
pylint
